@@ -10,7 +10,7 @@ import Foundation
 @available(iOS 13.0, watchOS 6.0, tvOS 13.0, macOS 10.15, *)
 extension YouTube {
     
-    public enum ExtractionMethod: Hashable {
+    public enum ExtractionMethod: Hashable, Sendable {
         case local
         case remote(serverURL: URL)
         
