@@ -7,13 +7,17 @@
 
 import Foundation
 
-class ConfigurableURLSessionDelegate: NSObject, URLSessionTaskDelegate {
+final class ConfigurableURLSessionDelegate: NSObject, URLSessionTaskDelegate {
     
     let allowsRedirects: Bool
     let applyCookiesOnRedirect: Bool
     let saveIntermediateResponses: Bool
     
+#if swift(>=5.10)
+    nonisolated(unsafe) var intermediateResponses = [HTTPURLResponse]()
+#else
     var intermediateResponses = [HTTPURLResponse]()
+#endif
     
     init(allowsRedirect: Bool, applyCookiesOnRedirect: Bool, saveIntermediateResponses: Bool) {
         self.allowsRedirects = allowsRedirect
